p61 Admin edit Product Gallery

+++ app/Http/Livewire/Admin/AdminProductComponent.php
@@ -20,8 +20,11 @@ class AdminProductComponent extends Component
         {
             $images = explode(",",$product->images);
             foreach($images as $image)
-            {
-                unlink('assets/images/products'.'/'.$image);
+            {   
+                if($image)
+                {
+                    unlink('assets/images/products'.'/'.$image);
+                }                
             }
         }
         $product->delete();
